+++ lib/yt/models/playlist.rb
@@ -157,13 +157,13 @@ module Yt
       has_report :views, Integer
 
       # @macro report_by_playlist_dimensions
-      has_report :estimated_minutes_watched, Float
+      has_report :estimated_minutes_watched, Integer
 
       # @macro report_by_gender_and_age_group
       has_report :viewer_percentage, Float
 
       # @macro report_by_day_and_state
-      has_report :average_view_duration, Float
+      has_report :average_view_duration, Integer
 
       # @macro report_by_day_and_state
       has_report :playlist_starts, Integer
